requirements.txt
fastapi
uvicorn[standard]
trafilatura
requests
pydantic
lxml
lxml_html_clean
python-dotenv
beautifulsoup4
slowapi
PyPDF2
python-multipart